+++ finagle-core/src/test/java/com/twitter/finagle/ssl/client/SslClientConfigurationCompilationTest.java
@@ -21,6 +21,7 @@ public class SslClientConfigurationCompilationTest {
 
     SslClientConfiguration config = new SslClientConfiguration(
       hostname,
+      hostname,
       KeyCredentialsConfig.UNSPECIFIED,
       TrustCredentialsConfig.UNSPECIFIED,
       CipherSuitesConfig.UNSPECIFIED,